Revert "Feat-22 / Pipeline Refactorization"
