--- archives/baseline/model.py
@@ -1,249 +1,0 @@
-import math
-
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-
-from loss import EASTLoss
-
-
-cfg = [64, 64, 'M', 128, 128, 'M', 256, 256, 256, 'M', 512, 512, 512, 'M', 512, 512, 512, 'M']
-
-
-def make_layers(cfg, batch_norm=False):
-    layers = []
-    in_channels = 3
-    for v in cfg:
-        if v == 'M':
-            layers += [nn.MaxPool2d(kernel_size=2, stride=2)]
-        else:
-            conv2d = nn.Conv2d(in_channels, v, kernel_size=3, padding=1)
-            if batch_norm:
-                layers += [conv2d, nn.BatchNorm2d(v), nn.ReLU(inplace=True)]
-            else:
-                layers += [conv2d, nn.ReLU(inplace=True)]
-            in_channels = v
-    return nn.Sequential(*layers)
-
-
-class VGG(nn.Module):
-    def __init__(self, features):
-        super(VGG, self).__init__()
-        self.features = features
-        self.avgpool = nn.AdaptiveAvgPool2d((7, 7))
-        self.classifier = nn.Sequential(
-            nn.Linear(512 * 7 * 7, 4096),
-            nn.ReLU(True),
-            nn.Dropout(),
-            nn.Linear(4096, 4096),
-            nn.ReLU(True),
-            nn.Dropout(),
-            nn.Linear(4096, 1000),
-        )
-
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                nn.init.kaiming_normal_(m.weight, mode='fan_out', nonlinearity='relu')
-                if m.bias is not None:
-                    nn.init.constant_(m.bias, 0)
-            elif isinstance(m, nn.BatchNorm2d):
-                nn.init.constant_(m.weight, 1)
-                nn.init.constant_(m.bias, 0)
-            elif isinstance(m, nn.Linear):
-                nn.init.normal_(m.weight, 0, 0.01)
-                nn.init.constant_(m.bias, 0)
-
-    def forward(self, x):
-        x = self.features(x)
-        x = self.avgpool(x)
-        x = x.view(x.size(0), -1)
-        x = self.classifier(x)
-        return x
-
-
-class Extractor(nn.Module):
-    def __init__(self, pretrained):
-        super().__init__()
-        vgg16_bn = VGG(make_layers(cfg, batch_norm=True))
-        if pretrained:
-            vgg16_bn.load_state_dict(torch.load('/data/ephemeral/home/.GitHub/pths/vgg16_bn-6c64b313.pth'))
-        self.features = vgg16_bn.features
-
-    def forward(self, x):
-        out = []
-        for m in self.features:
-            x = m(x)
-            if isinstance(m, nn.MaxPool2d):
-                out.append(x)
-        return out
-
-
-class Merge(nn.Module):
-    def __init__(self):
-        super().__init__()
-
-        # self.conv1 = nn.Conv2d(1024, 128, 1)
-        # self.bn1 = nn.BatchNorm2d(128)
-        # self.relu1 = nn.ReLU()
-        # self.conv2 = nn.Conv2d(128, 128, 3, padding=1)
-        # self.bn2 = nn.BatchNorm2d(128)
-        # self.relu2 = nn.ReLU()
-
-        # self.conv3 = nn.Conv2d(384, 64, 1)
-        # self.bn3 = nn.BatchNorm2d(64)
-        # self.relu3 = nn.ReLU()
-        # self.conv4 = nn.Conv2d(64, 64, 3, padding=1)
-        # self.bn4 = nn.BatchNorm2d(64)
-        # self.relu4 = nn.ReLU()
-
-        # self.conv5 = nn.Conv2d(192, 32, 1)
-        # self.bn5 = nn.BatchNorm2d(32)
-        # self.relu5 = nn.ReLU()
-        # self.conv6 = nn.Conv2d(32, 32, 3, padding=1)
-        # self.bn6 = nn.BatchNorm2d(32)
-        # self.relu6 = nn.ReLU()
-
-        # self.conv7 = nn.Conv2d(32, 32, 3, padding=1)
-        # self.bn7 = nn.BatchNorm2d(32)
-        # self.relu7 = nn.ReLU()
-
-        self.conv1 = nn.Conv2d(1024, 128, 1)
-        self.bn1 = nn.BatchNorm2d(128)
-        self.relu1 = nn.ReLU()
-        self.conv2 = nn.Conv2d(128, 128, 3, padding=1)
-        self.bn2 = nn.BatchNorm2d(128)
-        self.relu2 = nn.ReLU()
-
-        self.conv3 = nn.Conv2d(384, 64, 1)
-        self.bn3 = nn.BatchNorm2d(64)
-        self.relu3 = nn.ReLU()
-        self.conv4 = nn.Conv2d(64, 64, 3, padding=1)
-        self.bn4 = nn.BatchNorm2d(64)
-        self.relu4 = nn.ReLU()
-
-        self.conv5 = nn.Conv2d(192, 32, 1)
-        self.bn5 = nn.BatchNorm2d(32)
-        self.relu5 = nn.ReLU()
-        self.conv6 = nn.Conv2d(32, 32, 3, padding=1)
-        self.bn6 = nn.BatchNorm2d(32)
-        self.relu6 = nn.ReLU()
-
-        self.conv7 = nn.Conv2d(96, 32, 1)
-        self.bn7 = nn.BatchNorm2d(32)
-        self.relu7 = nn.ReLU()
-        self.conv8 = nn.Conv2d(32, 32, 3, padding=1)
-        self.bn8 = nn.BatchNorm2d(32)
-        self.relu8 = nn.ReLU()
-
-        self.conv9 = nn.Conv2d(32, 32, 3, padding=1)
-        self.bn9 = nn.BatchNorm2d(32)
-        self.relu9 = nn.ReLU()
-
-
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                nn.init.kaiming_normal_(m.weight, mode='fan_out', nonlinearity='relu')
-                if m.bias is not None:
-                    nn.init.constant_(m.bias, 0)
-            elif isinstance(m, nn.BatchNorm2d):
-                nn.init.constant_(m.weight, 1)
-                nn.init.constant_(m.bias, 0)
-
-    def forward(self, x):
-        # y = F.interpolate(x[3], scale_factor=2, mode='bilinear', align_corners=True)
-        # y = torch.cat((y, x[2]), 1)
-        # y = self.relu1(self.bn1(self.conv1(y)))
-        # y = self.relu2(self.bn2(self.conv2(y)))
-
-        # y = F.interpolate(y, scale_factor=2, mode='bilinear', align_corners=True)
-        # y = torch.cat((y, x[1]), 1)
-        # y = self.relu3(self.bn3(self.conv3(y)))
-        # y = self.relu4(self.bn4(self.conv4(y)))
-
-        # y = F.interpolate(y, scale_factor=2, mode='bilinear', align_corners=True)
-        # y = torch.cat((y, x[0]), 1)
-        # y = self.relu5(self.bn5(self.conv5(y)))
-        # y = self.relu6(self.bn6(self.conv6(y)))
-
-        # y = self.relu7(self.bn7(self.conv7(y)))
-
-        y = F.interpolate(x[4], scale_factor=2, mode='bilinear', align_corners=True)
-        y = torch.cat((y, x[3]), 1)
-        y = self.relu1(self.bn1(self.conv1(y)))
-        y = self.relu2(self.bn2(self.conv2(y)))
-
-        y = F.interpolate(y, scale_factor=2, mode='bilinear', align_corners=True)
-        y = torch.cat((y, x[2]), 1)
-        y = self.relu3(self.bn3(self.conv3(y)))
-        y = self.relu4(self.bn4(self.conv4(y)))
-
-        y = F.interpolate(y, scale_factor=2, mode='bilinear', align_corners=True)
-        y = torch.cat((y, x[1]), 1)
-        y = self.relu5(self.bn5(self.conv5(y)))
-        y = self.relu6(self.bn6(self.conv6(y)))
-
-        y = F.interpolate(y, scale_factor=2, mode='bilinear', align_corners=True)
-        y = torch.cat((y, x[0]), 1)
-        y = self.relu7(self.bn7(self.conv7(y)))
-        y = self.relu8(self.bn8(self.conv8(y)))
-
-        y = self.relu9(self.bn9(self.conv9(y)))
-
-        return y
-
-
-class Output(nn.Module):
-    def __init__(self, scope=512):
-        super().__init__()
-        self.conv1 = nn.Conv2d(32, 1, 1)
-        self.sigmoid1 = nn.Sigmoid()
-        self.conv2 = nn.Conv2d(32, 4, 1)
-        self.sigmoid2 = nn.Sigmoid()
-        self.conv3 = nn.Conv2d(32, 1, 1)
-        self.sigmoid3 = nn.Sigmoid()
-        self.scope = 512
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                nn.init.kaiming_normal_(m.weight, mode='fan_out', nonlinearity='relu')
-                if m.bias is not None:
-                    nn.init.constant_(m.bias, 0)
-
-    def forward(self, x):
-        score = self.sigmoid1(self.conv1(x))
-        loc   = self.sigmoid2(self.conv2(x)) * self.scope
-        angle = (self.sigmoid3(self.conv3(x)) - 0.5) * math.pi
-        geo   = torch.cat((loc, angle), 1)
-        return score, geo
-
-
-class EAST(nn.Module):
-    def __init__(self, pretrained=True):
-        super(EAST, self).__init__()
-        self.extractor = Extractor(pretrained)
-        self.merge = Merge()
-        self.output = Output()
-
-        self.criterion = EASTLoss()
-
-    def forward(self, x):
-        return self.output(self.merge(self.extractor(x)))
-
-    def train_step(self, image, score_map, geo_map, roi_mask):
-        device = list(self.parameters())[0].device
-        image, score_map, geo_map, roi_mask = (image.to(device), score_map.to(device),
-                                               geo_map.to(device), roi_mask.to(device))
-        pred_score_map, pred_geo_map = self.forward(image)
-
-        loss, values_dict = self.criterion(score_map, pred_score_map, geo_map, pred_geo_map,
-                                           roi_mask)
-        extra_info = dict(**values_dict, score_map=pred_score_map, geo_map=pred_geo_map)
-
-        return loss, extra_info
-
-
-if __name__ == '__main__':
-    m = EAST()
-    x = torch.randn(1, 3, 1024, 1024)
-    score, geo = m(x)
-    print(score.shape)
-    print(geo.shape)
