+++ baseline/detect.py
@@ -131,7 +131,7 @@ def detect(model, images, input_size, map_scale=0.5):
     for score_map, geo_map, orig_size in zip(score_maps, geo_maps, orig_sizes):
         map_margin = int(abs(orig_size[0] - orig_size[1]) * map_scale * input_size / max(orig_size))
         if orig_size[0] == orig_size[1]:
-            score_map, geo_map = score_map, geo_map
+            score_map, geo_map = score_map, geo_map        
         elif orig_size[0] > orig_size[1]:
             score_map, geo_map = score_map[:, :, :-map_margin], geo_map[:, :, :-map_margin]
         else:
