+++ requirements.txt
@@ -1,0 +1,7 @@
+lanms==1.0.2
+numpy==1.24.4
+opencv-python==4.8.1.78
+shapely==2.0.2
+tqdm==4.66.1
+matplotlib
+albumentations==1.2.0
