+++ train.py
@@ -9,208 +9,227 @@ from argparse import ArgumentParser
 
 import torch
 from torch import cuda
-from torch.utils.data import DataLoader
-from tqdm import tqdm
-import wandb
+from torch.utils.data import DataLoader, random_split
+from torch.optim import lr_scheduler
 
-# Function from "baseline" folder (수정이 불가한 내용들)
 from baseline.east_dataset import EASTDataset
+from dataset import SceneTextDataset
 from baseline.model import EAST
-from baseline.loss import EASTLoss
+from utils import increment_path
 
-# Funtction from code (수정이 가능한 내용들)
-from deteval import calc_deteval_metrics
-from utils import get_gt_bboxes, get_pred_bboxes, seed_everything, AverageMeter
-from dataset import SceneTextDataset, PickleDataset
-from optimizer import optim
-from scheduler import sched
-
-import albumentations as A
+import wandb
 import numpy as np
-os.environ['SM_MODEL_DIR'] = '/data/ephemeral/home/github'
+
 
 def parse_args():
     parser = ArgumentParser()
 
-    # 여러 피클 데이터셋 경로
-    parser.add_argument('--train_dataset_dirs', nargs='+', type=str, default=[
-        "/data/ephemeral/home/data/chinese_receipt/pickle/[1024, 1536, 2048]_cs[1024]_aug['CJ', 'GB', 'HSV', 'N']/train",
-        "/data/ephemeral/home/data/japanese_receipt/pickle/[1024, 1536, 2048]_cs[1024]_aug['CJ', 'GB', 'HSV', 'N']/train",
-        "/data/ephemeral/home/data/thai_receipt/pickle/[1024, 1536, 2048]_cs[1024]_aug['CJ', 'GB', 'HSV', 'N']/train",
-        "/data/ephemeral/home/data/vietnamese_receipt/pickle/[1024, 1536, 2048]_cs[1024]_aug['CJ', 'GB', 'HSV', 'N']/train"
-    ])
-    parser.add_argument('--data_dirs', nargs='+', type=str, default=[
-        "/data/ephemeral/home/data/chinese_receipt",
-        "/data/ephemeral/home/data/japanese_receipt",
-        "/data/ephemeral/home/data/thai_receipt",
-        "/data/ephemeral/home/data/vietnamese_receipt"
-    ])
-    parser.add_argument('--model_dir', type=str, default=os.environ.get('SM_MODEL_DIR', 'trained_models'))
-    parser.add_argument('--seed', type=int, default=4096)
-    parser.add_argument('--fold', type=int, default=0)
-    parser.add_argument('--val_interval', type=int, default=5)
-    parser.add_argument('--device', default='cuda:0' if cuda.is_available() else 'cpu')
-    parser.add_argument('--num_workers', type=int, default=8)
-    parser.add_argument('--image_size', type=int, default=2048)
-    parser.add_argument('--input_size', type=int, default=1024)
-    parser.add_argument('--batch_size', type=int, default=4)
-    parser.add_argument('--learning_rate', type=float, default=1e-3)
-    parser.add_argument('--max_epoch', type=int, default=150)
-    parser.add_argument('--save_interval', type=int, default=1)
-    parser.add_argument('--ignore_tags', type=list, default=['masked', 'excluded-region', 'maintable', 'stamp'])
-    parser.add_argument('-m', '--mode', type=str, default='on', help='wandb logging mode(on: online, off: disabled)')
-    parser.add_argument('-p', '--project', type=str, default='datacentric', help='wandb project name')
-    parser.add_argument('-d', '--data', default='pickle', type=str, help='description about dataset', choices=['original', 'pickle'])
-    parser.add_argument("--optimizer", type=str, default='Adam', choices=['adam', 'adamW'])
-    parser.add_argument("--scheduler", type=str, default='cosine', choices=['multistep', 'cosine'])
-    parser.add_argument("--resume", type=str, default=None, choices=[None, 'resume', 'finetune'])
-    parser.add_argument('--no-color_jitter', action='store_false', dest='color_jitter', default=True, help='Disable color jitter augmentation (default: True)')
-    parser.add_argument('--no-normalize', action='store_false', dest='normalize', default=True, help='Disable normalization (default: True)')
-    parser.add_argument('--apply_flip', action='store_true', help='Apply horizontal flip (default: False)')
-    parser.add_argument('--apply_rotate', action='store_true', help='Apply random rotate 90 (default: False)')
-    parser.add_argument('--apply_blur', action='store_true', help='Apply Gaussian blur (default: False)')
-    parser.add_argument('--save_dir', type=str, default=os.path.join(os.environ.get('SM_MODEL_DIR', 'trained_models'), 'saved_models'),
-                        help='Directory to save models')
+    # Conventional args
+    parser.add_argument("--seed", type=int, default=4669)
+    parser.add_argument("--data_dir", type=str, default="../data")
+    parser.add_argument("--model_dir", type=str, default="trained_models")
+    parser.add_argument("--device", default="cuda" if cuda.is_available() else "cpu")
+    parser.add_argument("--num_workers", type=int, default=8)
+    parser.add_argument("--image_size", type=int, default=2048)
+    parser.add_argument("--input_size", type=int, default=1024)
+    parser.add_argument("--batch_size", type=int, default=8)
+    parser.add_argument("--patience", type=int, default=10)
+    parser.add_argument("--learning_rate", type=float, default=1e-3)
+    parser.add_argument("--max_epochs", type=int, default=150)
+    parser.add_argument("--save_interval", type=int, default=5)
+    parser.add_argument("--ignore_tags", type=list, default=["masked", "excluded-region", "maintable", "stamp"])
+    parser.add_argument("--wandb_name", type=str, default="default_run_name")
+    parser.add_argument("--validation_split", type=float, default=0.2)  # Validation split 비율 추가
+    parser.add_argument("--augmentation", type=int, default=0)
+    parser.add_argument("--binarization", type=int, default=0)
+    parser.add_argument("--color_jitter", type=int, default=0)
+    parser.add_argument("--normalize", type=int, default=0)
+
     args = parser.parse_args()
 
     if args.input_size % 32 != 0:
-        raise ValueError('`input_size` must be a multiple of 32')
-
+        raise ValueError("`input_size` must be a multiple of 32")
     return args
 
 
-def create_transforms(args):
-    funcs = []
-
-    if args.color_jitter:
-        funcs.append(A.ColorJitter(brightness=0.2, contrast=0.2, saturation=0.2, hue=0.1))
-
-    if args.apply_flip:
-        funcs.append(A.HorizontalFlip(p=0.5))
-
-    if args.apply_rotate:
-        funcs.append(A.RandomRotate90(p=0.5))
-
-    if args.apply_blur:
-        funcs.append(A.GaussianBlur(blur_limit=(3, 7), p=0.3))
-
-    if args.normalize:
-        funcs.append(A.Normalize(mean=(0.6831708235495132, 0.6570838514500981, 0.6245893701608299),
-                                 std=(0.19835448743425943, 0.20532970462804873, 0.21117810051894778)))
-
-    transform = A.Compose(funcs)
-    return transform
+def seed_everything(seed):
+    torch.manual_seed(seed)
+    torch.cuda.manual_seed(seed)
+    torch.cuda.manual_seed_all(seed)  # if using multi-GPU
+    torch.backends.cudnn.deterministic = True
+    torch.backends.cudnn.benchmark = False
+    np.random.seed(seed)
+    random.seed(seed)
+    os.environ["PYTHONHASHSEED"] = str(seed)
 
 
-def do_training(args):
-    for data_dir, train_dataset_dir in zip(args.data_dirs, args.train_dataset_dirs):
-        dataset_name = osp.basename(data_dir)  # 데이터셋 이름 추출
-        save_dir = osp.join(args.save_dir, dataset_name)  # 데이터셋별 저장 경로 생성
-        os.makedirs(save_dir, exist_ok=True)
+def do_training(config, seed, dataset_path, model_dir, device, image_size, input_size, num_workers, batch_size,
+                patience, learning_rate, max_epochs, save_interval, ignore_tags, wandb_name, validation_split, augmentation, binarization,
+                color_jitter, normalize):
+    # 시드 설정
+    if seed == -1:
+        seed = int.from_bytes(os.urandom(4), byteorder="big")
+    print(f"seed: {seed}")
+    seed_everything(seed)
 
-        transform = create_transforms(args)
-        model = EAST().to(args.device)
-        optimizer = optim(args.optimizer, args.learning_rate, model.parameters())
-        scheduler = sched(args, optimizer)
+    # 모델 디렉토리 설정
+    model_dir = increment_path(os.path.join(model_dir, wandb_name))
+    os.makedirs(model_dir, exist_ok=True)
+    model_name = osp.basename(model_dir)
 
-        # WandB 초기화 및 watch
-        if args.mode == 'on':
-            wandb.init(
-                project=args.project,
-                entity='cv-21',
-                group=osp.basename(data_dir),
-                name=f'{args.max_epoch}e_{args.optimizer}_{args.scheduler}_{args.learning_rate}_{dataset_name}'
+    # config.json 로그 파일 저장
+    with open(os.path.join(model_dir, f"{model_name}.json"), "w", encoding="utf-8") as f:
+        json.dump(vars(config), f, ensure_ascii=False, indent=4)
+
+    # wandb 로깅 설정
+    wandb.init(project="Lv.2_data_centric", name=model_name, config=config)
+
+    # 데이터셋 초기화
+    full_dataset = SceneTextDataset(
+        dataset_path,
+        split="train",
+        image_size=image_size,
+        crop_size=input_size,
+        ignore_tags=ignore_tags,
+        augmentation=augmentation,
+        binarization=binarization,
+        color_jitter=color_jitter,
+        normalize=normalize
+    )
+    full_dataset = EASTDataset(full_dataset)
+
+    # 학습 및 검증 데이터셋 분할
+    train_size = int((1 - validation_split) * len(full_dataset))
+    valid_size = len(full_dataset) - train_size
+    train_dataset, valid_dataset = random_split(full_dataset, [train_size, valid_size])
+
+    train_num_batches = math.ceil(len(train_dataset) / batch_size)
+    valid_num_batches = math.ceil(len(valid_dataset) / batch_size)
+
+    # 데이터 로더 설정
+    train_loader = DataLoader(
+        train_dataset,
+        batch_size=batch_size,
+        shuffle=True,
+        num_workers=num_workers
+    )
+    valid_loader = DataLoader(
+        valid_dataset,
+        batch_size=batch_size,
+        shuffle=False,
+        num_workers=num_workers
+    )
+
+    # 모델, 옵티마이저, 스케줄러 설정
+    model = EAST().to(device)
+    optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
+    scheduler = lr_scheduler.MultiStepLR(optimizer, milestones=[max_epochs // 2], gamma=0.1)
+
+    # Early stopping 초기화
+    counter = 0
+    best_val_loss = np.inf
+
+    # ========== 학습 시작 ==========
+    for epoch in range(max_epochs):
+        model.train()
+        train_loss, valid_loss, train_start = 0, 0, time.time()
+        for idx, (img, gt_score_map, gt_geo_map, roi_mask) in enumerate(train_loader):
+            loss, extra_info = model.train_step(img, gt_score_map, gt_geo_map, roi_mask)
+            optimizer.zero_grad()
+            loss.backward()
+            optimizer.step()
+
+            loss_val = loss.item()
+            train_loss += loss_val
+
+            # CLI 로그
+            print(
+                f"Epoch[{epoch+1:02}/{max_epochs}]({idx+1:02}/{len(train_loader)}) || "
+                f"Learning Rate: {scheduler.get_last_lr()[0]} || "
+                f"Train Loss: {loss_val:4.4f} || "
+                f"Train Class loss: {extra_info['cls_loss']:4.4f} || "
+                f"Train Angle loss: {extra_info['angle_loss']:4.4f} || "
+                f"Train IoU loss: {extra_info['iou_loss']:4.4f}"
             )
-            wandb.config.update(args)
-            wandb.watch(model)
 
-        if args.data == 'pickle':
-            train_dataset = PickleDataset(train_dataset_dir)
-        else:
-            train_dataset = SceneTextDataset(
-                data_dir,
-                split='train',
-                json_name=f'train{args.fold}.json',
-                image_size=args.image_size,
-                crop_size=args.input_size,
-                ignore_tags=args.ignore_tags,
-                pin_memory=True,
-            )
-            train_dataset = EASTDataset(train_dataset)
+            # wandb 로그
+            wandb.log({
+                "Train Cls loss": extra_info['cls_loss'],
+                "Train Angle loss": extra_info['angle_loss'],
+                "Train IoU loss": extra_info['iou_loss'],
+                "Train Loss": loss_val,
+                "Learning Rate": scheduler.get_last_lr()[0],
+                "Epoch": epoch + 1,
+                "Seed": seed
+            })
 
-        train_num_batches = math.ceil(len(train_dataset) / args.batch_size)
-        train_loader = DataLoader(
-            train_dataset,
-            batch_size=args.batch_size,
-            shuffle=True,
-            num_workers=args.num_workers,
-            pin_memory=True
-        )
+        train_end = time.time() - train_start
 
-        ### Val Loader ###
-        valid_json_file = f'ufo/valid{args.fold}.json'
-        val_images = []
-        if osp.exists(osp.join(data_dir, valid_json_file)):
-            with open(osp.join(data_dir, valid_json_file), 'r', encoding='utf-8') as file:
-                val_data = json.load(file)
-            val_images = list(val_data['images'].keys())
+        # ========== 검증 ==========
+        model.eval()
+        with torch.no_grad():
+            valid_start = time.time()
+            for idx, (img, gt_score_map, gt_geo_map, roi_mask) in enumerate(valid_loader):
+                loss, extra_info = model.train_step(img, gt_score_map, gt_geo_map, roi_mask)
+                valid_loss += loss.item()
 
-        best_f1_score = 0
-        for epoch in range(args.max_epoch):
-            model.train()
-            with tqdm(total=train_num_batches) as pbar:
-                for img, gt_score_map, gt_geo_map, roi_mask in train_loader:
-                    img = [image.cpu().numpy().transpose(1, 2, 0) if isinstance(image, torch.Tensor) else image for image in img]
-                    img = [transform(image=image)['image'] for image in img]
-                    img = [torch.from_numpy(image).permute(2, 0, 1) for image in img]
-                    img = torch.stack(img)
-                    loss, extra_info = model.train_step(img, gt_score_map, gt_geo_map, roi_mask)
-                    optimizer.zero_grad()
-                    loss.backward()
-                    optimizer.step()
+            mean_val_loss = valid_loss / valid_num_batches
+            if best_val_loss > mean_val_loss:
+                best_val_loss = mean_val_loss
+                ckpt_fpath = osp.join(model_dir, f"best_epoch_{epoch+1}.pth")
+                torch.save(model.state_dict(), ckpt_fpath)
+                counter = 0
+            else:
+                counter += 1
 
-                    train_dict = {
-                        'train total loss': loss.item(),
-                        'train cls loss': extra_info['cls_loss'],
-                        'train angle loss': extra_info['angle_loss'],
-                        'train iou loss': extra_info['iou_loss']
-                    }
-                    pbar.update(1)
-                    pbar.set_postfix(train_dict)
-                    if args.mode == 'on':
-                        wandb.log(train_dict, step=epoch)
+        valid_end = time.time() - valid_start
+        print(f"Valid Mean loss: {mean_val_loss:.4f}")
 
-            scheduler.step()
+        if counter > patience:
+            print("Early stopping triggered.")
+            break
 
-            if (epoch + 1) % args.val_interval == 0 or epoch >= args.max_epoch - 5:
-                print("Calculating validation results...")
-                pred_bboxes_dict = get_pred_bboxes(model, data_dir, val_images, args.input_size, args.batch_size, split='train')
-                gt_bboxes_dict = get_gt_bboxes(data_dir, json_file=valid_json_file, valid_images=val_images)
-                result = calc_deteval_metrics(pred_bboxes_dict, gt_bboxes_dict)
-                precision, recall = result['total']['precision'], result['total']['recall']
-                f1_score = 2*precision*recall/(precision+recall) if precision + recall > 0 else 0
-                print(f'Precision: {precision} Recall: {recall} F1 Score: {f1_score}')
-
-                val_dict = {'val precision': precision, 'val recall': recall, 'val f1_score': f1_score}
-                if args.mode == 'on':
-                    wandb.log(val_dict, step=epoch)
-
-                if f1_score > best_f1_score:
-                    best_f1_score = f1_score
-                    torch.save(model.state_dict(), osp.join(save_dir, 'best.pth'))
-
-            if (epoch + 1) % args.save_interval == 0:
-                torch.save(model.state_dict(), osp.join(save_dir, 'latest.pth'))
-
-        if args.mode == 'on':
-            wandb.alert('Training Task Finished', f"Best F1 Score: {best_f1_score:.4f}")
-            wandb.finish()
+        scheduler.step()
 
 
 def main(args):
-    do_training(args)
+    languages = ["chinese_receipt", "japanese_receipt", "thai_receipt", "vietnamese_receipt"]
 
-if __name__ == '__main__':
+    for language in languages:
+        print(f"Starting training for {language}")
+        dataset_path = osp.join(args.data_dir, language)
+
+        # 언어별 저장 경로 지정
+        language_save_dir = osp.join(args.save_dir, language)
+        os.makedirs(language_save_dir, exist_ok=True)
+
+        # wandb 이름을 언어별로 지정하여 구분 가능하도록 설정
+        args.wandb_name = f"{language}_run"
+
+        # 각 언어 데이터셋에 대해 학습 진행
+        do_training(
+            config=args,
+            seed=args.seed,
+            dataset_path=dataset_path,
+            model_dir=language_save_dir,
+            device=args.device,
+            image_size=args.image_size,
+            input_size=args.input_size,
+            num_workers=args.num_workers,
+            batch_size=args.batch_size,
+            patience=args.patience,
+            learning_rate=args.learning_rate,
+            max_epochs=args.max_epochs,
+            save_interval=args.save_interval,
+            ignore_tags=args.ignore_tags,
+            wandb_name=args.wandb_name,
+            validation_split=args.validation_split,
+            augmentation=args.augmentation,
+            binarization=args.binarization,
+            color_jitter=args.color_jitter,
+            normalize=args.normalize
+        )
+
+if __name__ == "__main__":
     args = parse_args()
-    seed_everything(args.seed)
     main(args)
